Undo Update AutoReset.cpp

+++ src/Player/AutoReset.cpp
@@ -4,8 +4,8 @@
 class $modify(PlayLayer) {
     void updateProgressbar() {
 		PlayLayer::updateProgressbar();	
-		float percent = (m_player1->getPositionX() / m_levelLength) * 100;
-		if (((float)percent == getVar<float>("auto_reset_num")) && getVar<bool>("auto_reset")) {
+		double percent = (m_player1->getPositionX() / m_levelLength) * 100;
+		if (((int)percent == getVar<float>("auto_reset_num")) && getVar<bool>("auto_reset")) {
 			resetLevel();
 		}
     }
